+++ E_coli_disaccharides_minimal_monomers.csv
