+++ E_coli_disaccharides_monomers_only.csv
