+++ OT_200124_Ecoli_WT_1_Rep1_Cleaned_10ppm_tol.xlsx
